refactor(openapi): confine SupportedAppType to the openapi layer
Move SupportedAppType / SUPPORTED_APP_TYPES out of the shared models.model
into controllers/openapi/_models.py, where they are the only consumers
(query type + _is_listable post-filter). The CLI whitelist still derives
from the same generated zSupportedAppType — codegen is unchanged because
swagger names the enum by its field type regardless of definition site.

Also drops changes that had no benefit once filtering is done by
post-filter + query-type validation, reverting them to main:
- app_service.py: mode param type + filter-ladder rewrite (kept original
  Literal + != AGENT denylist for the 'all' case)
- console/app/app.py caller and roster.py mode arg
- the agent-service filter test

Fixes the type-check-core failure (console Literal mode vs AppListParams
AppMode|Literal['all']) by keeping app_service on its original Literal.

The PR now touches only the openapi controllers, openapi docs/contracts,
the CLI, and their tests.
