+++ api/services/app_service.py
@@ -45,7 +45,7 @@ AppListSortBy = Literal["last_modified", "recently_created", "earliest_created"]
 class AppListBaseParams(BaseModel):
     page: int = Field(default=1, ge=1)
     limit: int = Field(default=20, ge=1, le=100)
-    mode: AppMode | Literal["all"] = "all"
+    mode: Literal["completion", "chat", "advanced-chat", "workflow", "agent-chat", "agent", "channel", "all"] = "all"
     sort_by: AppListSortBy = "last_modified"
     name: str | None = None
     tag_ids: list[str] | None = None
@@ -84,10 +84,20 @@ class AppService:
     ) -> list[sa.ColumnElement[bool]]:
         filters = [App.tenant_id == tenant_id, App.is_universal == False]
 
-        if params.mode == "all":
+        if params.mode == "workflow":
+            filters.append(App.mode == AppMode.WORKFLOW)
+        elif params.mode == "completion":
+            filters.append(App.mode == AppMode.COMPLETION)
+        elif params.mode == "chat":
+            filters.append(App.mode == AppMode.CHAT)
+        elif params.mode == "advanced-chat":
+            filters.append(App.mode == AppMode.ADVANCED_CHAT)
+        elif params.mode == "agent-chat":
+            filters.append(App.mode == AppMode.AGENT_CHAT)
+        elif params.mode == "agent":
+            filters.append(App.mode == AppMode.AGENT)
+        elif params.mode == "all":
             filters.append(App.mode != AppMode.AGENT)
-        else:
-            filters.append(App.mode == params.mode)
 
         if isinstance(params, AppListParams):
             if params.status:
