+++ api/tests/unit_tests/models/test_app_models.py
@@ -20,7 +20,6 @@ from sqlalchemy.dialects import postgresql
 
 from models.enums import ConversationFromSource
 from models.model import (
-    SUPPORTED_APP_TYPES,
     App,
     AppAnnotationHitHistory,
     AppAnnotationSetting,
@@ -31,7 +30,6 @@ from models.model import (
     Message,
     MessageAnnotation,
     Site,
-    SupportedAppType,
     load_annotation_reply_config,
 )
 
@@ -117,21 +115,6 @@ class TestAppModelValidation:
 
         with pytest.raises(ValueError, match="invalid mode value"):
             AppMode.value_of("invalid_mode")
-
-    def test_supported_app_type_is_the_listable_subset_of_app_mode(self):
-        """SupportedAppType (and the derived SUPPORTED_APP_TYPES tuple) is exactly the
-        curated, listable subset of AppMode; non-app/runtime modes stay out."""
-        assert {t.value for t in SupportedAppType} == {
-            "completion",
-            "chat",
-            "advanced-chat",
-            "workflow",
-            "agent-chat",
-        }
-        assert set(SUPPORTED_APP_TYPES) <= set(AppMode)
-        assert AppMode.AGENT not in SUPPORTED_APP_TYPES
-        assert AppMode.RAG_PIPELINE not in SUPPORTED_APP_TYPES
-        assert AppMode.CHANNEL not in SUPPORTED_APP_TYPES
 
     def test_icon_type_validation(self):
         """Test icon type enum values."""
